[ALPS07175600] DISP: mt6789 solve Coverity issue
[Detail]
mt6789 solve Coverity issue, PRINTF_ARGS.

MTK-Commit-Id: 0199056e23b516c888fdfd0b1f81586719ae2396

CR-Id: ALPS07175600
Feature: display_mobile

+++ drivers/gpu/drm/mediatek/mediatek_v2/mtk_dsi.c
@@ -6054,7 +6054,7 @@ unsigned long long mtk_dsi_get_frame_hrt_bw_base_by_datarate(
 		if (vact)
 			bw_base = bw_base * vtotal / vact;
 		else
-			bw_base = vtotal * hact * vrefresh * 4 / 1000;
+			bw_base = (unsigned long long)vtotal * hact * vrefresh * 4 / 1000;
 		bw_base = bw_base / 1000;
 	} else {
 		bw_base = data_rate * dsi->lanes * compress_rate * 4;
